[ADD] Add entities for SDL Library + Change vector to Int

+++ Core/include/arcade/Geometry.hh
@@ -13,19 +13,19 @@ namespace arc {
 
     class Vector {
         public:
-            Vector(float pos_x, float pos_y);
+            Vector(int pos_x, int pos_y);
             Vector(const Vector&);
             ~Vector();
             Vector &operator=(const Vector&);
 
-            [[nodiscard]] float getVectorX() const;
-            [[nodiscard]] float getVectorY() const;
-            void setVectorX(float new_point);
-            void setVectorY(float new_point);
+            [[nodiscard]] int getVectorX() const;
+            [[nodiscard]] int getVectorY() const;
+            void setVectorX(int new_point);
+            void setVectorY(int new_point);
 
         private:
-            float x;
-            float y;
+            int x;
+            int y;
     };
 
     class Rect {
